fix notes loading path and move component's file location
